+++ source/Examples/Resources/xSccmPreReqs/SccmPreReqs.ps1
@@ -25,8 +25,6 @@ Configuration Example
             InstallMdt             = $true
             InstallWindowsFeatures = $true
             SccmRole               = 'CASorSiteServer'
-            AddWindowsFirewallRule = $true
-            FirewallProfile        = 'Domain','Private'
             AdkSetupExePath        = 'C:\temp\ADKInstall\adksetup.exe'
             AdkWinPeSetupPath      = 'C:\temp\ADKInstall\adkwinpesetup.exe'
             MdtMsiPath             = 'C:\temp\MDTInstall\MicrosoftDeploymentToolkit_x64_1809.msi'